Support renaming kernelspecs at runtime.
This change adds support for kernel spec managers that rename kernel specs
based on configured traits.

This is a necessary step in the work to support multiplexing between
multiple kernel spec managers (#1187), as we need to be able to rename
kernel specs in order to prevent collisions between the kernel specs
provided by multiple kernel spec managers.